move draggable components to its own component
